requirements.txt
#requirements.txt


# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# MongoDB
motor==3.3.2
pymongo==4.6.0

# GraphQL
strawberry-graphql[fastapi]==0.217.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Storage & CDN
boto3==1.29.7

# Redis for caching
redis==5.0.1

# Add this line to requirements.txt
sendgrid==6.10.0

# Image processing
pillow==10.1.0

# HTTP client
httpx==0.25.2

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Utils
python-dateutil==2.8.2
pytz==2023.3